requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# RAG and document processing
langchain==0.0.335
langchain-openai==0.0.2
chromadb==0.4.18
pypdf==3.17.1
faiss-cpu==1.7.4
tiktoken==0.5.1

# Speech processing
openai==1.3.5
azure-cognitiveservices-speech==1.31.0

# Wake word detection
pvporcupine==2.2.1

# Security and utilities
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.3
httpx==0.25.1